adding some range checking

+++ Sockets/ServerForAss3.php
@@ -10,6 +10,8 @@ $socket = socket_create(AF_INET, SOCK_STREAM, 0);
 $result = socket_bind($socket, $host, $port) or die("Could not bind port, please try again later\n");
 // start listening for connections
 $result = socket_listen($socket, 3);
+$lowerRange = 0; //later we will get the range by joining the node network and requesting a node number
+$higherRange = 255;
 
 echo trim(shell_exec('hostname'))."\n";
 echo "Port: $port\n";
@@ -23,53 +25,65 @@ while(1)
     $input = socket_read($spawn, 1024) or die("Could not read input\n");
     $input = trim($input);
    
+
+
     $command = intval(substr($input,0,2),16);
     $key = intval(substr($input, 2,64),16);
-
-    if($command == 1) 
-    { //PUT operation
-        
-        $valueLength = intval(substr($input,66,4),16);
-        $value = substr($input,70,$valueLength);
-        
-        if(file_put_contents($key,$value) == false) 
-        {
-            $response = '02'; //out of space response
-        } 
-        else 
-        {
-            $response = '00'; //operation successful response
-        }
-
-    } 
-    elseif ($command == 2) 
-    { //GET operation
-        
-        $filecontents = file_get_contents($key);
-        
-        if($filecontents != false) {
-            $response = '00'.strlen($filecontents).$filecontents; //operation successful + value length + value
-        } 
-        else 
-        {
-            $response = '01'; //key not found response
-        }
-
-    } 
-    elseif ($command == 3) 
-    { //REMOVE operation
-        
-        if(unlink($key) == true) 
-        {
-            $response = '00'; //operation successful response
-        } 
-        else 
-        {
-            $response = '01'; //key not found response
-        }
+    $hashKey = substr(hash('md5',$key),0,2); //hash the key, get the first two characters, which represent an int between 0 and 255
     
+    if(($lowerRange < $upperRange and $lowerRange > intval($hashKey,16) and $upperRange =< intval($hashKey,16)) 
+        or ($lowerRange > $upperRange and ($lowerRange < intval($hashKey,16) or $upperRange => intval($hashKey,16))) 
+    //check if the value is in the range serviced by this node
+    {
+
+        
+
+        if($command == 1) 
+        { //PUT operation
+            
+            $valueLength = intval(substr($input,66,4),16);
+            $value = substr($input,70,$valueLength);
+            
+            if(file_put_contents($key,$value) == false) 
+            {
+                $response = '02'; //out of space response
+            } 
+            else 
+            {
+                $response = '00'; //operation successful response
+            }
+
+        } 
+        elseif ($command == 2) 
+        { //GET operation
+            
+            $filecontents = file_get_contents($key);
+            
+            if($filecontents != false) {
+                $response = '00'.strlen($filecontents).$filecontents; //operation successful + value length + value
+            } 
+            else 
+            {
+                $response = '01'; //key not found response
+            }
+
+        } 
+        elseif ($command == 3) 
+        { //REMOVE operation
+            
+            if(unlink($key) == true) 
+            {
+                $response = '00'; //operation successful response
+            } 
+            else 
+            {
+                $response = '01'; //key not found response
+            }
+        
+        }
+    } else {
+        //pass query to node in successor list
     }
-
